requirements.txt
crewai>=0.11.0
anthropic>=0.25.0
langchain>=0.1.13
langchain-openai>=0.1.0
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.5.0
python-dotenv>=1.0.0
requests>=2.31.0
sqlalchemy>=2.0.0
pytest>=7.4.0
pytest-cov>=4.1.0
pyyaml>=6.0.1
httpx>=0.25.0
aiosqlite>=0.22.0
